requirements.txt
icoextract==0.1.6
numpy==2.2.5
Pillow==11.2.1
pygame==2.6.1
PyOpenGL==3.1.9
pywin32==310